fix(orders): typos, datatypes and queries

+++ src/main/java/ar/edu/ungs/fleet_manager/orders/application/OrderRequest.java
@@ -1,7 +1,9 @@
 package ar.edu.ungs.fleet_manager.orders.application;
 
+import java.math.BigDecimal;
+
 public record OrderRequest (String providerId,
                             String productId,
                             Integer quantity,
-                            Integer amount){
+                            BigDecimal amount){
 }
